Parallel work i main

+++ Consid23/ConsoleOnlySubmitter.cs
@@ -21,14 +21,17 @@ public class ConsoleOnlySubmitter : ISolutionSubmitter
     public void AddSolutionToSubmit(SubmitSolution sol)
     {
         var score = _scorer.CalculateScore(sol);
-        Console.WriteLine($"Not submitting (ConsoleOnlySubmitter) GameScore: {score.GameScore.Total} co2 {score.GameScore.KgCo2Savings * _generalData.Co2PricePerKiloInSek} earnings {score.GameScore.Earnings} footfall {score.GameScore.TotalFootfall}");
 
+        lock(this)
         if (score.GameScore.Total > _maxSubmitted)
         {
+            Console.WriteLine($"Not submitting (ConsoleOnlySubmitter) GameScore: {score.GameScore.Total} co2 {score.GameScore.KgCo2Savings * _generalData.Co2PricePerKiloInSek} earnings {score.GameScore.Earnings} footfall {score.GameScore.TotalFootfall}");
+
             _maxSubmitted = score.GameScore.Total;
             var time = DateTime.Now.ToString("dd_hh_mm_ss");
             var filename = $"{_mapData.MapName}_{score.GameScore.Total}_{time}.json";
-            File.WriteAllText(filename, JsonConvert.SerializeObject(sol));
+            if(File.Exists(filename) == false)
+                File.WriteAllText(filename, JsonConvert.SerializeObject(sol));
         }
     }
